feat: game object gen, part 7, data-driven most blocks

+++ script/Pack.py
@@ -1,6 +1,7 @@
 import os
 import tempfile
 import shutil
+from zipfile import ZipFile
 from pathlib import Path
 
 from CreativeTabProvider import CreativeTabProvider
@@ -20,6 +21,13 @@ class Pack:
             self.config['translation_file_name'] = '{}.json'
         if 'translation_dest' in self.config:
             self.config['translation_dest'] = self.config['translation_dest'].replace('%TEMP%', self.tempDir)
+        self.includes = []
+        if 'includes' in self.config:
+            self.includes = self.config['includes']
+        for include in self.includes:
+            path = Path(include)
+            if not path.is_dir():
+                raise ValueError('Include path is not a directory: ' + path.absolute().as_posix())
         self.creativeTabs = CreativeTabProvider(self)
         self.translations = TranslationProvider(self)
         print('Temp directory:', self.tempDir)
@@ -34,12 +42,25 @@ class Pack:
             provider.generate()
             print('Generated', provider.count, 'files for', str(provider))
         self.providers.clear()
+
+        for include in self.includes:
+            shutil.copytree(include, self.tempDir, dirs_exist_ok=True)
+
         dest = Path(self.config['dest'])
-        if dest.exists():
-            shutil.rmtree(dest)
-        dest.mkdir(parents=True, exist_ok=True)
-        for f in os.listdir(self.tempDir):
-            shutil.move(os.path.join(self.tempDir, f), dest)
+        if dest.name.lower().endswith('.zip'):
+            if not Path(os.path.join(self.tempDir, 'pack.mcmeta')).exists():
+                raise ValueError('pack.mcmeta not found in dest package')
+            dest.parent.mkdir(parents=True, exist_ok=True)
+            with ZipFile(dest, 'w') as zip:
+                for root, _, files in os.walk(self.tempDir):
+                    for file in files:
+                        zip.write(os.path.join(root, file), os.path.relpath(os.path.join(root, file), self.tempDir))
+        else:
+            if dest.exists():
+                shutil.rmtree(dest)
+            dest.mkdir(parents=True, exist_ok=True)
+            for f in os.listdir(self.tempDir):
+                shutil.move(os.path.join(self.tempDir, f), dest)
         shutil.rmtree(self.tempDir)
 
     def defaultResourceLocation(self, path: str) -> ResourceLocation:
